writer/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.12.3
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
google-ai-generativelanguage==0.6.6
    # via google-generativeai
google-api-core==2.19.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.140.0
    # via google-generativeai
google-auth==2.33.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.65.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
proto-plus==1.24.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.4
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via google-generativeai
pydantic-core==2.20.1
    # via pydantic
pyparsing==3.1.2
    # via httplib2
python-dotenv==1.0.1
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via google-auth
soupsieve==2.5
    # via beautifulsoup4
tqdm==4.66.5
    # via google-generativeai
typing-extensions==4.12.2
    # via
    #   google-generativeai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
